requirements.txt
fastapi>=0.116.0,<1.0.0
oracledb>=3.3.0,<4.0.0
PyYAML>=6.0.2,<7.0.0
uvicorn[standard]>=0.35.0,<1.0.0
